requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
crispy-bootstrap4==2024.10
dj-database-url==2.1.0
Django==4.2.20
django-bootstrap-datepicker-plus==5.0.4
django-crispy-forms==2.4
django-filter==23.3
mysql==0.0.3
mysql-connector-python==8.1.0
mysqlclient==2.2.7
pillow==11.2.1
protobuf==4.21.12
pydantic==2.11.3
pydantic_core==2.33.1
sqlparse==0.5.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
